Move application delegate to its own file
