+++ src/platform_impl/ios/mod.rs
@@ -58,6 +58,7 @@
 #![cfg(ios_platform)]
 #![allow(clippy::let_unit_value)]
 
+mod app_delegate;
 mod app_state;
 mod event_loop;
 mod ffi;
